catalog-react: moved DefaultStarredEntitiesApi implementation to catalog plugin
