+++ plugins/catalog/package.json
@@ -44,6 +44,7 @@
     "@backstage/plugin-catalog-react": "^0.7.0",
     "@backstage/search-common": "^0.2.4",
     "@backstage/theme": "^0.2.15",
+    "@backstage/types": "^0.1.2",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "4.0.0-alpha.57",
@@ -51,7 +52,8 @@
     "lodash": "^4.17.21",
     "react-helmet": "6.1.0",
     "react-router": "6.0.0-beta.0",
-    "react-use": "^17.2.4"
+    "react-use": "^17.2.4",
+    "zen-observable": "^0.8.15"
   },
   "peerDependencies": {
     "@types/react": "^16.13.1 || ^17.0.0",
