+++ plugins/catalog/src/apis/StarredEntitiesApi/DefaultStarredEntitiesApi.test.ts
