+++ plugins/catalog/src/apis/StarredEntitiesApi/migration.test.ts
