+++ plugins/catalog/src/apis/StarredEntitiesApi/migration.ts
