+++ plugins/catalog/src/index.ts
@@ -20,6 +20,8 @@
  * @packageDocumentation
  */
 
+export * from './apis';
+
 export * from './components/AboutCard';
 export * from './components/CatalogKindHeader';
 export * from './components/CatalogSearchResultListItem';
